Move filters to a separated component
