+++ web/src/services/loans.service.js
@@ -157,7 +157,7 @@ const getFundingLoans = () => {
     })
     .catch((error) => {
       console.log(error);
-      return false;
+      return [];
     });
 };
